Testing and fine tuning
Program is almost complete, some things are missing, as have been pointed out by the comments in main.c

+++ Arduino/DigitalTeknikProjekt/DigitalTeknikProjekt/Globals.c
@@ -1,4 +1,4 @@
 #include "Globals.h"
 
-int packetLength = 0;
+unsigned long packetLength = 0;
 int compareValue = 256;
